Skip already existing download

+++ main.py
@@ -22,4 +22,7 @@ if __name__ == "__main__":
         print(f"Downloading {target['timestamp']} for {url}")
         file = Path(argv[2]) / target["timestamp"] / "index.html"
         file.parent.mkdir(parents=True, exist_ok=True)
+        if file.exists():
+            print(f"File {file} already exists, skipping download.")
+            continue
         file.write_bytes(download_website(url, target["timestamp"], proxy))
